controlplane: remove obsolete CNPNodeStatusGC test
[ upstream commit 51101e9afc38d1db7c12404fc4629fe7f66a7586 ]

[ backporter notes: Fixes conflict in
  test/controlplane/ciliumnetworkpolicies/status_updates_gc.go ]

The control plane test for CNPNodeStatusGC was initially intended to
check the behaviour of running with and without GC. With the deprecation
of the CNP node status, the GC now amounts to unconditional cleanup.

The commit mentioned below had broken the test insofar as it remove the
differentiating factor: the configuration. In other words, we were
running the same code twice, expecting a different outcome.

Ideally, this would have broken the first variant - with GC "disabled",
but unfortunately the test was racy in itself. I suspect that with
progressing modularisation of the agent and operator, somewhere along
the line we lost the guarantee that the GC happens before StartOperator
returns. Hence we checked that the CNPs were unchanged while the
operator was starting up concurrently - a classic race condition.

Since the whole test is obsolete, simply remove it, and leave the second
variant of the test around to check that we actually perform the
deletion.

Fixes: c15f8e4a24 (Remove skip-cnp-status-startup-clean)

Co-authored-by: Fabian Fischer <fabian.fischer@isovalent.com>
Signed-off-by: David Bimmler <david.bimmler@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>