ClusterMesh: validate etcd cluster ID for remote connections
In a configuration where there are mutliple replicas of the
clustermesh-apiserver, each Pod runs its own etcd instance with a unique
cluster ID. This commit adds a `clusterLock` type, which is a wrapper
around a unit64 that can only be set once, that is used to create a gRPC
unary and stream interceptor. The interceptors are provided to the etcd
client to intercept and validate the cluster ID in the header of all
responses from the etcd server.

If the client receives a response from a different cluster, the
connection is terminated and restarted. This is necessary to prevent
resuming watches or relying on the status check from another etcd
cluster.

Signed-off-by: Tim Horner <timothy.horner@isovalent.com>