hive/cell/health: don't warn when reporting on stopped reporter.
[ upstream commit 44e90054f29c410608e816b112df8fc615eb0131 ]

Currently in cases where reporters emit status on stopped reporters, a warning is logged.
This causes failures in CI related to endpoints, as those close their reporters after endpoint is deleted.

This scenario is probably not a valuable log for users and generally such occurrences should be harmless.
This moves the log to debug level.

Fixes: #31147

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>