ci: Fix renovate datasource for ClusterLoader2
The commit for ClusterLoader2 was not being updated by renovate due to
an incorrect configuration. This commit adjusts the configuration to
ensure that CL2 is automatically updated by renovate. It also manually
updates the commit to point to the current ref in the master branch at
the time of writing, March 14th, 2024.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>