wip: Allow overriding identity from UpsertMetadata
Given the CI failures regarding:

```
Found 52 k8s-app=cilium logs matching list of errors that must be investigated:
    2024-03-16T07:47:32.097814910Z level=warning msg="Detected identity override, but no labels where specified. Falling back on the old non-override labels. This may cause connectivity issues for this address." cidr="fd02::1cc/128" oldIdentity= resource=cep/kube-system/coredns-76f75df574-dqpmv subsys=ipcache
```

my theory is that the separate calls to OverrideIdentity() and
UpsertMetadata() inside the CEP watcher were causing a race where the
metadata resolution was processing the identity override attribute, but
before the labels have been processed. Therefore, insert the identity
override attribute along with the labels together in one call in attempt
to eliminate this race.

This commit should be squashed with the relevant commit if CI improves
with this commit.

Signed-off-by: Chris Tarazi <chris@isovalent.com>